+++ ENIGMAsystem/SHELL/Bridges/Cocoa-OpenGL/graphics_bridge.cpp
@@ -31,8 +31,6 @@ extern "C" {
 }
 
 namespace enigma {
-  GLuint msaa_fbo = 0;
-
   extern void (*WindowResizedCallback)();
   void WindowResized() {
     // clear the window color, viewport does not need set because backbuffer was just recreated
@@ -54,10 +52,5 @@ namespace enigma {
 namespace enigma_user {
   void set_synchronization(bool enable) {
 
-  }
-
-  void display_reset(int samples, bool vsync) {
-    set_synchronization(vsync);
-    //TODO: Copy over from the Win32 bridge
   }
 }
