+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL/GLscreen.h
@@ -1,0 +1,7 @@
+
+namespace enigma {
+
+void gl_screen_init();
+void msaa_fbo_blit();
+
+} // namespace enigma
