+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL/Makefile
@@ -1,0 +1,10 @@
+SOURCES += $(wildcard Graphics_Systems/OpenGL/*.cpp)
+
+ifeq ($(TARGET-PLATFORM), Windows)
+	override CXXFLAGS += -DGLEW_STATIC
+	override LDLIBS += -lglew32 -lopengl32
+else ifeq ($(TARGET-PLATFORM), Linux)
+	override LDLIBS += -lGL -lGLEW
+else ifeq ($(TARGET-PLATFORM), MacOSX)
+	override LDLIBS += -framework OpenGL -lGLEW
+endif
