+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL1/GLscreen.cpp
@@ -18,6 +18,7 @@
 **/
 
 #include "Graphics_Systems/graphics_mandatory.h"
+#include "Graphics_Systems/OpenGL/GLscreen.h"
 #include "Graphics_Systems/OpenGL/OpenGLHeaders.h"
 #include "Graphics_Systems/General/GSscreen.h"
 #include "Graphics_Systems/General/GSbackground.h"
@@ -44,7 +45,6 @@ using namespace enigma;
 
 namespace enigma {
 
-extern GLuint msaa_fbo;
 unsigned int bound_framebuffer = 0; //Shows the bound framebuffer, so glGetIntegerv(GL_FRAMEBUFFER_BINDING_EXT, &fbo); don't need to be called (they are very slow)
 int viewport_x, viewport_y, viewport_w, viewport_h; //These are used by surfaces, to set back the viewport
 
@@ -53,7 +53,7 @@ void scene_begin() {
 }
 
 void scene_end() {
-
+	msaa_fbo_blit();
 }
 
 } // namespace enigma
